refactor: add MapLayerGroup component
